🎨[structure] 파일 분리
