+++ client/src/assets/SearchIcon.tsx
@@ -10,7 +10,7 @@ const SearchIcon = () => {
     >
       <path
         d="M13.5 1.6875C20.9554 1.6875 27 6.59812 27 12.6562C27 18.7144 20.9554 23.625 13.5 23.625C12.7845 23.625 12.0808 23.5794 11.3957 23.4917C8.49656 26.3908 5.03381 26.9122 1.6875 26.9882V26.2777C3.49481 25.3918 5.0625 23.7802 5.0625 21.9375C5.0625 21.681 5.04225 21.4279 5.00513 21.1815C1.95244 19.17 0 16.0987 0 12.6562C0 6.59812 6.04462 1.6875 13.5 1.6875Z"
-        fill="#BBBBBB"
+        fill="#FFFFFF"
       />
     </svg>
   );
